CC | image: move image service related code into image-rpc.rs
Move image service related code into image-rpc.rs, to simplify
maintenance.
